+++ shopGunSdk/src/main/java/com/shopgun/android/sdk/ShopGun.java
@@ -25,9 +25,6 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
 
-import com.apollographql.apollo.ApolloClient;
-import com.apollographql.apollo.api.ScalarType;
-import com.apollographql.apollo.response.CustomTypeAdapter;
 import com.shopgun.android.sdk.api.Environment;
 import com.shopgun.android.sdk.api.EnvironmentEvents;
 import com.shopgun.android.sdk.api.EnvironmentGraph;
@@ -139,8 +136,6 @@ public class ShopGun {
     private String mSessionId;
     /** The device id, this will if possible be persisted across installations */
     private String mDeviceId;
-    /** Apollo client for graphQL queries */
-    private ApolloClient mApolloClient;
 
     // Things we'd like to get rid of
 
@@ -174,7 +169,6 @@ public class ShopGun {
         mExecutor = builder.executorService;
         mClient = builder.okHttpClient;
         mRealmConfiguration = builder.realmConfiguration;
-        mApolloClient = builder.apolloClient;
         mLegacyConfiguration = builder.legacyConfiguration;
 
         mLifecycleManager = new LifecycleManager(builder.application);
@@ -228,16 +222,6 @@ public class ShopGun {
      */
     public OkHttpClient getClient() {
         return mClient;
-    }
-
-    /**
-     * Get the client for graphQL queries to
-     * one of the {@link EnvironmentGraph}
-     * Currently we're using Apollo library to perform graph queries.
-     * @return ApolloClient
-     */
-    public ApolloClient getGraphClient() {
-        return mApolloClient;
     }
 
     public LifecycleManager getLifecycleManager() {
@@ -502,10 +486,8 @@ public class ShopGun {
         RealmConfiguration realmConfiguration;
         RealmConfiguration legacyConfiguration;
         OkHttpClient okHttpClient;
-        ApolloClient apolloClient;
         List<Interceptor> interceptors = new ArrayList<>();
         List<Interceptor> networkInterceptors = new ArrayList<>();
-        SimpleArrayMap<ScalarType, CustomTypeAdapter<?>> apolloCustomTypeAdapters = new SimpleArrayMap<>();
 
         /**
          * Start building your {@link ShopGun} instance.
@@ -525,11 +507,6 @@ public class ShopGun {
 
         public Builder addNetworkInterceptor(Interceptor interceptor) {
             networkInterceptors.add(interceptor);
-            return this;
-        }
-
-        public Builder addApolloCustomTypeAdapter(ScalarType scalarType, CustomTypeAdapter<?> adapter) {
-            apolloCustomTypeAdapters.put(scalarType, adapter);
             return this;
         }
 
@@ -706,16 +683,6 @@ public class ShopGun {
             // Add sdk interceptors last to override user options if necessary
             okHttpClientBuilder.addInterceptor(new UserAgentInterceptor(SgnUserAgent.getUserAgent(application)));
             okHttpClient = okHttpClientBuilder.build();
-
-            // Setup the default ApolloClient
-            ApolloClient.Builder apolloBuilder = ApolloClient.builder()
-                    .serverUrl(graphEnvironment.toString())
-                    .okHttpClient(okHttpClient);
-            for (int i = 0; i < apolloCustomTypeAdapters.size(); i++) {
-                apolloBuilder.addCustomTypeAdapter(
-                        apolloCustomTypeAdapters.keyAt(i), apolloCustomTypeAdapters.valueAt(i));
-            }
-            apolloClient = apolloBuilder.build();
 
             // Set the default RealmConfiguration.
             Realm.init(application);